frontend/requirements.txt
attrs==25.1.0
beautifulsoup4==4.13.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
mysql-connector-python==9.2.0
outcome==1.3.0.post0
packaging==24.2
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.28.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.37
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
Flask==2.0.1
Werkzeug==2.0.3

# python -m venv env
# pip install -r env\requirements.txt
# .\env\Scripts\Activate.ps1
# cd env
# python.exe projetos.py; python.exe portarias.py; python.exe agenda_camara.py; python.exe agenda_presidentes.py; python.exe agenda_ministro_saude.py; python.exe eventos_deputados.py

